Refactor code, add 2018 form back and create seperate test files for 2024 version
