+++ lib/pdf_fill/forms/va214142.rb
@@ -30,13 +30,13 @@ module PdfFill
         },
         'veteranSocialSecurityNumber' => {
           'first' => {
-            key: 'F[0].Page_1[0].VeteransSocialSecurityNumber_FirstThreeNumbers[0]'
+            key: 'F[0].Page_1[0].ClaimantsSocialSecurityNumber_FirstThreeNumbers[0]'
           },
           'second' => {
-            key: 'F[0].Page_1[0].VeteransSocialSecurityNumber_SecondTwoNumbers[0]'
+            key: 'F[0].Page_1[0].ClaimantsSocialSecurityNumber_SecondTwoNumbers[0]'
           },
           'third' => {
-            key: 'F[0].Page_1[0].VeteransSocialSecurityNumber_LastFourNumbers[0]'
+            key: 'F[0].Page_1[0].ClaimantsSocialSecurityNumber_LastFourNumbers[0]'
           }
         },
         'vaFileNumber' => {
@@ -54,68 +54,54 @@ module PdfFill
           }
         },
         'veteranServiceNumber' => {
-          key: 'F[0].Page_1[0].VeteransServiceNumber_If_Applicable[0]'
+          key: 'F[0].Page_1[0].VeteransServiceNumber[0]'
         },
         'veteranAddress' => {
           question_num: 6,
           question_text: 'MAILING ADDRESS',
 
           'street' => {
-            key: 'F[0].Page_1[0].MailingAddress_NumberAndStreet[0]',
+            key: 'F[0].Page_1[0].CurrentMailingAddress_NumberAndStreet[0]',
             limit: 30,
             question_num: 6,
             question_suffix: 'A',
             question_text: 'Number and Street'
           },
-          # TODO: Confirm that extra page is created for "See additional" when apt/unit number is used
           'street2' => {
-            key: 'F[0].Page_1[0].MailingAddress_ApartmentOrUnitNumber[0]',
+            key: 'F[0].Page_1[0].CurrentMailingAddress_ApartmentOrUnitNumber[0]',
             limit: 5,
             question_num: 6,
             question_suffix: 'B',
             question_text: 'Apartment or Unit Number'
           },
           'city' => {
-            key: 'F[0].Page_1[0].MailingAddress_City[0]',
+            key: 'F[0].Page_1[0].CurrentMailingAddress_City[0]',
             limit: 18,
             question_num: 6,
             question_suffix: 'C',
             question_text: 'City'
           },
           'state' => {
-            key: 'F[0].Page_1[0].MailingAddress_StateOrProvince[0]'
+            key: 'F[0].Page_1[0].CurrentMailingAddress_StateOrProvince[0]'
           },
           'country' => {
-            key: 'F[0].Page_1[0].MailingAddress_Country[0]',
+            key: 'F[0].Page_1[0].CurrentMailingAddress_Country[0]',
             limit: 2
           },
           'postalCode' => {
             'firstFive' => {
-              key: 'F[0].Page_1[0].MailingAddress_ZIPOrPostalCode_FirstFiveNumbers[0]'
+              key: 'F[0].Page_1[0].CurrentMailingAddress_ZIPOrPostalCode_FirstFiveNumbers[0]'
             },
             'lastFour' => {
-              key: 'F[0].Page_1[0].MailingAddress_ZIPOrPostalCode_LastFourNumbers[0]'
+              key: 'F[0].Page_1[0].CurrentMailingAddress_ZIPOrPostalCode_LastFourNumbers[0]'
             }
           }
         },
-        # TODO: 2018 form had one field for phone number, the 2024 form has 3 fields. Need to create a object similar to social security.
-        'veteranPhone' => {
-          'phone_area_code' => {
-            key: 'F[0].Page_1[0].TelephoneNumber_AreaCode[0]'
-          },
-          'phone_first_three_numbers' => {
-            key: 'F[0].Page_1[0].TelephoneNumber_SecondThreeNumbers[0]'
-          },
-          'phone_last_four_numbers' => {
-            key: 'F[0].Page_1[0].TelephoneNumber_LastFourNumbers[0]'
-          }
-        },
-        # TODO
-        'internationalPhoneNumber' => {
-          key: 'F[0].Page_1[0].International_Telephone_Number_If_Applicable[0]'
-        },
         'email' => {
-          key: 'F[0].Page_1[0].E_Mail_Address[0]'
+          key: 'F[0].Page_1[0].EMAIL[0]'
+        },
+        'veteranPhone' => {
+          key: 'F[0].Page_1[0].EMAIL[1]'
         },
         'veteranSocialSecurityNumber1' => {
           'first' => {
@@ -134,52 +120,28 @@ module PdfFill
           question_text: 'Limited Consent',
           key: 'F[0].#subform[1].InformationIsLimitedToWhatIsWrittenInThisSpace[0]'
         },
-        # NOTE: Signature field is not populating F[0].#subform[1].SignatureField11[0]
         'signature' => {
-          key: 'F[0].#subform[1].SignatureField11[0]'
+          key: 'F[0].#subform[1].CLAIMANT_SIGNATURE[0]'
         },
-        # TODO: 2018 form had one field for signature date, the 2024 form has 3 fields. Need to create a object similar to social security.
         'signatureDate' => {
-          'month' => {
-            key: 'F[0].#subform[1].Date_Signed_Month[0]'
-          },
-          'day' => {
-            key: 'F[0].#subform[1].Date_Signed_Day[0]'
-          },
-          'year' => {
-            key: 'F[0].#subform[1].Date_Signed_Year[0]'
-          }
+          key: 'F[0].#subform[1].DateSigned_Month_Day_Year[0]',
+          format: 'date'
         },
-        # TODO: 2018 form had one field for printed name, the 2024 form has 3 fields. Need to create a object similar to social security.
         'printedName' => {
-          'first' => {
-            key: 'F[0].#subform[1].Printed_Name_Of_Person_Signing_First[0]',
-            limit: 12,
-            question_num: 1,
-            question_text: "VETERAN/BENEFICIARY'S FIRST NAME"
-          },
-          'middleInitial' => {
-            key: 'F[0].#subform[1].Printed_Name_Of_Person_Signing_Middle_Initial[0]'
-          },
-          'last' => {
-            key: 'F[0].#subform[1].Printed_Name_Of_Person_Signing_Last[0]',
-            limit: 18,
-            question_num: 1,
-            question_text: "VETERAN/BENEFICIARY'S LAST NAME"
-          }
+          key: 'F[0].#subform[1].PrintedNameOfPersonAuthorizingDisclosure[0]'
         },
         'veteranFullName1' => {
           'first' => {
-            key: 'F[0].#subform[14].VeteranFirstName[0]',
+            key: 'F[0].#subform[8].VeteranFirstName[0]',
             limit: 12,
             question_num: 17,
             question_text: "4142a VETERAN/BENEFICIARY'S FIRST NAME"
           },
           'middleInitial' => {
-            key: 'F[0].#subform[14].VeteranMiddleInitial1[0]'
+            key: 'F[0].#subform[8].VeteranMiddleInitial1[0]'
           },
           'last' => {
-            key: 'F[0].#subform[14].VeteranLastName[0]',
+            key: 'F[0].#subform[8].VeteranLastName[0]',
             limit: 18,
             question_num: 17,
             question_text: "4142a VETERAN/BENEFICIARY'S LAST NAME"
@@ -187,49 +149,93 @@ module PdfFill
         },
         'veteranSocialSecurityNumber2' => {
           'first' => {
-            key: 'F[0].#subform[14].SSN1[0]'
+            key: 'F[0].#subform[8].ClaimantsSocialSecurityNumber_FirstThreeNumbers[0]'
           },
           'second' => {
-            key: 'F[0].#subform[14].SSN2[0]'
+            key: 'F[0].#subform[8].ClaimantsSocialSecurityNumber_SecondTwoNumbers[0]'
           },
           'third' => {
-            key: 'F[0].#subform[14].SSN3[0]'
+            key: 'F[0].#subform[8].ClaimantsSocialSecurityNumber_LastFourNumbers[0]'
           }
         },
-        'veteranSocialSecurityNumber3' => {
-          'first' => {
-            key: 'F[0].#subform[15].SSN1[1]'
-          },
-          'second' => {
-            key: 'F[0].#subform[15].SSN2[1]'
-          },
-          'third' => {
-            key: 'F[0].#subform[15].SSN3[1]'
-          }
-        },
-        # NOTE: This is field is not populating F[0].#subform[14].VAFileNumber[0] is the new field in the 2024 form
         'vaFileNumber1' => {
-          key: 'F[0].#subform[14].VAFileNumber[0]'
+          key: 'F[0].#subform[8].VAFileNumber[0]'
         },
         'veteranDateOfBirth1' => {
           'month' => {
-            key: 'F[0].#subform[14].Month[0]'
+            key: 'F[0].#subform[8].DOBmonth[0]'
           },
           'day' => {
-            key: 'F[0].#subform[14].Day[0]'
+            key: 'F[0].#subform[8].DOBday[0]'
           },
           'year' => {
-            key: 'F[0].#subform[14].Year[0]'
+            key: 'F[0].#subform[8].DOByear[0]'
           }
         },
         'veteranServiceNumber1' => {
-          key: 'F[0].#subform[14].VeteransServiceNumber_If_Applicable[0]'
+          key: 'F[0].#subform[8].VeteransServiceNumber[0]'
+        },
+        'veteranSocialSecurityNumber3' => {
+          'first' => {
+            key: 'F[0].#subform[9].VeteransSocialSecurityNumber_FirstThreeNumbers[1]'
+          },
+          'second' => {
+            key: 'F[0].#subform[9].VeteransSocialSecurityNumber_SecondTwoNumbers[1]'
+          },
+          'third' => {
+            key: 'F[0].#subform[9].VeteransSocialSecurityNumber_LastFourNumbers[1]'
+          }
         },
         'providerFacility' => {
           limit: 5,
-          first_key: 'providerFacilityName_0',
+          first_key: 'providerFacilityName',
           question_text: 'PROVIDER / FACILITY',
 
+          'providerFacilityName' => {
+            key: "F[0].provider.name[#{PROVIDER_ITERATOR}]"
+          },
+          'dateRangeStart0' => {
+            key: "F[0].provider.dateRangeStart0[#{PROVIDER_ITERATOR}]",
+            format: 'date'
+          },
+          'dateRangeEnd0' => {
+            key: "F[0].provider.dateRangeEnd0[#{PROVIDER_ITERATOR}]",
+            format: 'date'
+          },
+          'dateRangeStart1' => {
+            key: "F[0].provider.dateRangeStart1[#{PROVIDER_ITERATOR}]",
+            format: 'date'
+          },
+          'dateRangeEnd1' => {
+            key: "F[0].provider.dateRangeEnd1[#{PROVIDER_ITERATOR}]",
+            format: 'date'
+          },
+          'street' => {
+            limit: 30,
+            key: "F[0].provider.numberAndStreet[#{PROVIDER_ITERATOR}]"
+          },
+          'street2' => {
+            limit: 5,
+            key: "F[0].provider.apartmentOrUnitNumber[#{PROVIDER_ITERATOR}]"
+          },
+          'city' => {
+            limit: 18,
+            key: "F[0].provider.city[#{PROVIDER_ITERATOR}]"
+          },
+          'state' => {
+            key: "F[0].provider.state[#{PROVIDER_ITERATOR}]"
+          },
+          'country' => {
+            key: "F[0].provider.country[#{PROVIDER_ITERATOR}]"
+          },
+          'postalCode' => {
+            'firstFive' => {
+              key: "F[0].provider.postalCode_FirstFiveNumbers[#{PROVIDER_ITERATOR}]"
+            },
+            'lastFour' => {
+              key: "F[0].provider.postalCode_LastFourNumbers[#{PROVIDER_ITERATOR}]"
+            }
+          },
           'nameAndAddressOfProvider' => {
             key: '',
             question_suffix: 'A',
@@ -241,185 +247,8 @@ module PdfFill
             question_suffix: 'B',
             question_text: 'Treatment Dates',
             question_num: 9
-          },
-
-          # Provider 0
-          'providerFacilityName_0' => { key: 'F[0].#subform[14].Provider_Or_Facility_Name[0]' },
-          'conditionsTreated_0' => { key: 'F[0].#subform[14].Conditions_You_Are_Being_Treated_For[0]' },
-          'dateRangeStart_0' => {
-            'month' => { key: 'F[0].#subform[14].Month[0]' },
-            'day' => { key: 'F[0].#subform[14].Day[0]' },
-            'year' => { key: 'F[0].#subform[14].Year[0]' }
-          },
-          'dateRangeEnd_0' => {
-            'month' => { key: 'F[0].#subform[14].Month[1]' },
-            'day' => { key: 'F[0].#subform[14].Day[1]' },
-            'year' => { key: 'F[0].#subform[14].Year[1]' }
-          },
-          'street_0' => { limit: 30, key: 'F[0].#subform[14].Provider_Facility_Street_Address_NumberAndStreet[0]' },
-          'street2_0' => { limit: 5, key: 'F[0].#subform[14].MailingAddress_ApartmentOrUnitNumber[0]' },
-          'city_0' => { limit: 18, key: 'F[0].#subform[14].Provider_Facility_Address_City[0]' },
-          'state_0' => { key: 'F[0].#subform[14].Provider_Facility_Address_StateOrProvince[0]' },
-          'country_0' => { key: 'F[0].#subform[14].Provider_Facility_Address_Country[0]' },
-          'postalCode_0' => {
-            'firstFive' => { key: 'F[0].#subform[14].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[0]' },
-            'lastFour' => { key: 'F[0].#subform[14].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[0]' }
-          },
-          # For additional pages
-          'nameAndAddressOfProvider_0' => {
-            key: '',
-            question_suffix: 'A',
-            question_text: 'Name and Address of Provider',
-            question_num: 9
-          },
-          'combinedTreatmentDates_0' => {
-            key: '',
-            question_suffix: 'B',
-            question_text: 'Treatment Dates',
-            question_num: 9
-          },
-
-          # Provider 1
-          'providerFacilityName_1' => { key: 'F[0].#subform[14].Provider_Or_Facility_Name[1]' },
-          'conditionsTreated_1' => { key: 'F[0].#subform[14].Conditions_You_Are_Being_Treated_For[1]' },
-          'dateRangeStart_1' => {
-            'month' => { key: 'F[0].#subform[14].Month[3]' },
-            'day' => { key: 'F[0].#subform[14].Day[3]' },
-            'year' => { key: 'F[0].#subform[14].Year[3]' }
-          },
-          'dateRangeEnd_1' => {
-            'month' => { key: 'F[0].#subform[14].Month[4]' },
-            'day' => { key: 'F[0].#subform[14].Day[4]' },
-            'year' => { key: 'F[0].#subform[14].Year[4]' }
-          },
-          'street_1' => { limit: 30, key: 'F[0].#subform[14].Provider_Facility_Street_Address_NumberAndStreet[1]' },
-          'street2_1' => { limit: 5, key: 'F[0].#subform[14].MailingAddress_ApartmentOrUnitNumber[1]' },
-          'city_1' => { limit: 18, key: 'F[0].#subform[14].Provider_Facility_Address_City[1]' },
-          'state_1' => { key: 'F[0].#subform[14].Provider_Facility_Address_StateOrProvince[1]' },
-          'country_1' => { key: 'F[0].#subform[14].Provider_Facility_Address_Country[1]' },
-          'postalCode_1' => {
-            'firstFive' => { key: 'F[0].#subform[14].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[1]' },
-            'lastFour' => { key: 'F[0].#subform[14].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[1]' }
-          },
-          'nameAndAddressOfProvider_1' => {
-            key: '',
-            question_suffix: 'A',
-            question_text: 'Name and Address of Provider',
-            question_num: 10
-          },
-          'combinedTreatmentDates_1' => {
-            key: '',
-            question_suffix: 'B',
-            question_text: 'Treatment Dates',
-            question_num: 10
-          },
-
-          # Provider 2
-          'providerFacilityName_2' => { key: 'F[0].#subform[15].Provider_Or_Facility_Name[2]' },
-          'conditionsTreated_2' => { key: 'F[0].#subform[15].Conditions_You_Are_Being_Treated_For[2]' },
-          'dateRangeStart_2' => {
-            'month' => { key: 'F[0].#subform[15].Month[5]' },
-            'day' => { key: 'F[0].#subform[15].Day[5]' },
-            'year' => { key: 'F[0].#subform[15].Year[5]' }
-          },
-          'dateRangeEnd_2' => {
-            'month' => { key: 'F[0].#subform[15].Month[6]' },
-            'day' => { key: 'F[0].#subform[15].Day[6]' },
-            'year' => { key: 'F[0].#subform[15].Year[6]' }
-          },
-          'street_2' => { limit: 30, key: 'F[0].#subform[15].Provider_Facility_Street_Address_NumberAndStreet[2]' },
-          'street2_2' => { limit: 5, key: 'F[0].#subform[15].MailingAddress_ApartmentOrUnitNumber[2]' },
-          'city_2' => { limit: 18, key: 'F[0].#subform[15].Provider_Facility_Address_City[2]' },
-          'state_2' => { key: 'F[0].#subform[15].Provider_Facility_Address_StateOrProvince[2]' },
-          'country_2' => { key: 'F[0].#subform[15].Provider_Facility_Address_Country[2]' },
-          'postalCode_2' => {
-            'firstFive' => { key: 'F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[2]' },
-            'lastFour' => { key: 'F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[2]' }
-          },
-          'nameAndAddressOfProvider_2' => {
-            key: '',
-            question_suffix: 'A',
-            question_text: 'Name and Address of Provider',
-            question_num: 11
-          },
-          'combinedTreatmentDates_2' => {
-            key: '',
-            question_suffix: 'B',
-            question_text: 'Treatment Dates',
-            question_num: 11
-          },
-
-          # Provider 3
-          'providerFacilityName_3' => { key: 'F[0].#subform[15].Provider_Or_Facility_Name[3]' },
-          'conditionsTreated_3' => { key: 'F[0].#subform[15].Conditions_You_Are_Being_Treated_For[3]' },
-          'dateRangeStart_3' => {
-            'month' => { key: 'F[0].#subform[15].Month[7]' },
-            'day' => { key: 'F[0].#subform[15].Day[7]' },
-            'year' => { key: 'F[0].#subform[15].Year[7]' }
-          },
-          'dateRangeEnd_3' => {
-            'month' => { key: 'F[0].#subform[15].Month[8]' },
-            'day' => { key: 'F[0].#subform[15].Day[8]' },
-            'year' => { key: 'F[0].#subform[15].Year[8]' }
-          },
-          'street_3' => { limit: 30, key: 'F[0].#subform[15].Provider_Facility_Street_Address_NumberAndStreet[3]' },
-          'street2_3' => { limit: 5, key: 'F[0].#subform[15].MailingAddress_ApartmentOrUnitNumber[3]' },
-          'city_3' => { limit: 18, key: 'F[0].#subform[15].Provider_Facility_Address_City[3]' },
-          'state_3' => { key: 'F[0].#subform[15].Provider_Facility_Address_StateOrProvince[3]' },
-          'country_3' => { key: 'F[0].#subform[15].Provider_Facility_Address_Country[3]' },
-          'postalCode_3' => {
-            'firstFive' => { key: 'F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[3]' },
-            'lastFour' => { key: 'F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[3]' }
-          },
-          'nameAndAddressOfProvider_3' => {
-            key: '',
-            question_suffix: 'A',
-            question_text: 'Name and Address of Provider',
-            question_num: 12
-          },
-          'combinedTreatmentDates_3' => {
-            key: '',
-            question_suffix: 'B',
-            question_text: 'Treatment Dates',
-            question_num: 12
-          },
-
-          # Provider 4
-          'providerFacilityName_4' => { key: 'F[0].#subform[15].Provider_Or_Facility_Name[4]' },
-          'conditionsTreated_4' => { key: 'F[0].#subform[15].Conditions_You_Are_Being_Treated_For[4]' },
-          'dateRangeStart_4' => {
-            'month' => { key: 'F[0].#subform[15].Month[9]' },
-            'day' => { key: 'F[0].#subform[15].Day[9]' },
-            'year' => { key: 'F[0].#subform[15].Year[9]' }
-          },
-          'dateRangeEnd_4' => {
-            'month' => { key: 'F[0].#subform[15].Month[10]' },
-            'day' => { key: 'F[0].#subform[15].Day[10]' },
-            'year' => { key: 'F[0].#subform[15].Year[10]' }
-          },
-          'street_4' => { limit: 30, key: 'F[0].#subform[15].Provider_Facility_Street_Address_NumberAndStreet[4]' },
-          'street2_4' => { limit: 5, key: 'F[0].#subform[15].MailingAddress_ApartmentOrUnitNumber[4]' },
-          'city_4' => { limit: 18, key: 'F[0].#subform[15].Provider_Facility_Address_City[4]' },
-          'state_4' => { key: 'F[0].#subform[15].Provider_Facility_Address_StateOrProvince[4]' },
-          'country_4' => { key: 'F[0].#subform[15].Provider_Facility_Address_Country[4]' },
-          'postalCode_4' => {
-            'firstFive' => { key: 'F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[4]' },
-            'lastFour' => { key: 'F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[4]' }
-          },
-          'nameAndAddressOfProvider_4' => {
-            key: '',
-            question_suffix: 'A',
-            question_text: 'Name and Address of Provider',
-            question_num: 13
-          },
-          'combinedTreatmentDates_4' => {
-            key: '',
-            question_suffix: 'B',
-            question_text: 'Treatment Dates',
-            question_num: 13
           }
         }
-
       }.freeze
 
       def expand_va_file_number
@@ -440,15 +269,6 @@ module PdfFill
         end
       end
 
-      def expand_phone_number
-        phone = @form_data['veteranPhone']
-        return if phone.blank?
-
-        ['', '1', '2', '3'].each do |suffix|
-          @form_data["veteranPhone#{suffix}"] = split_phone_number(phone)
-        end
-      end
-
       def expand_claimant_address
         @form_data['veteranAddress']['country'] = extract_country(@form_data['veteranAddress'])
         @form_data['veteranAddress']['postalCode'] = split_postal_code(@form_data['veteranAddress'])
@@ -460,12 +280,6 @@ module PdfFill
         end
       end
 
-      def expand_printed_full_name
-        ['', '1'].each do |suffix|
-          @form_data["printedName#{suffix}"] = extract_middle_i(@form_data, 'veteranFullName')
-        end
-      end
-
       def expand_veteran_dob
         veteran_date_of_birth = @form_data['veteranDateOfBirth']
         return if veteran_date_of_birth.blank?
@@ -473,14 +287,6 @@ module PdfFill
         ['', '1'].each do |suffix|
           @form_data["veteranDateOfBirth#{suffix}"] = split_date(veteran_date_of_birth)
         end
-      end
-
-      # TODO: refactor Date.strptime
-      def expand_signature_date
-        veteran_signature_date = Date.strptime(@form_data['signatureDate'], '%Y-%m-%d').to_s
-        return if veteran_signature_date.blank?
-
-        @form_data['signatureDate'] = split_date(veteran_signature_date)
       end
 
       def expand_veteran_service_number
@@ -494,45 +300,16 @@ module PdfFill
         end
       end
 
-      def expand_provider_facility_name(providers)
-        providers.each_with_index do |provider, i|
-          updates = {}
-
-          provider.each_key do |key| # key = providerFacilityName
-
-            next if key.match?(/nameAndAddressOfProvider|combinedTreatmentDates/)
-
-            if key == 'providerFacilityAddress'
-              updates["#{key}_#{i}"] = {}
-
-              provider['providerFacilityAddress'].each_key do |address_key| # address_key = street
-                updates["#{key}_#{i}"]["#{address_key}_#{i}"] =
-                  provider['providerFacilityAddress'][address_key]
-              end
-
-              provider.delete('providerFacilityAddress')
-            else
-              updates["#{key}_#{i}"] = provider[key]
-              provider.delete(key)
-            end
-          end
-
-          # updates
-          provider.merge!(updates)
-        end
-      end
-
       def expand_provider_date_range(providers)
         providers.each do |provider|
           dates_of_treatment = provider['treatmentDateRange']
           date_ranges = {}
-          dates_of_treatment.each do |date_range|
+          dates_of_treatment.each_with_index do |date_range, index|
             date_ranges.merge!(
-              'dateRangeStart' => split_date(date_range['from']),
-              'dateRangeEnd' => split_date(date_range['to'])
+              "dateRangeStart#{index}" => date_range['from'],
+              "dateRangeEnd#{index}" => date_range['to']
             )
           end
-
           provider.except!('treatmentDateRange')
           provider.merge!(date_ranges)
         end
@@ -554,18 +331,11 @@ module PdfFill
       end
 
       def expand_provider_extras(providers)
-binding.pry
-        providers.each_with_index do |provider, i|
+        providers.each do |provider|
           name_address_extras = combine_name_addr_extras(provider, 'providerFacilityName', 'providerFacilityAddress')
+          provider['nameAndAddressOfProvider'] = PdfFill::FormValue.new('', name_address_extras)
           dates_extras = combine_date_ranges(provider['treatmentDateRange'])
-
-          if providers.size > 5
-            provider['nameAndAddressOfProvider'] = PdfFill::FormValue.new('', name_address_extras)
-            provider['combinedTreatmentDates'] = PdfFill::FormValue.new('', dates_extras)
-          else
-            provider["nameAndAddressOfProvider_#{i}"] = PdfFill::FormValue.new('', name_address_extras)
-            provider["combinedTreatmentDates_#{i}"] = PdfFill::FormValue.new('', dates_extras)
-          end
+          provider['combinedTreatmentDates'] = PdfFill::FormValue.new('', dates_extras)
         end
       end
 
@@ -575,24 +345,17 @@ binding.pry
         expand_provider_extras(providers)
         expand_provider_address(providers)
         expand_provider_date_range(providers)
-
-        expand_provider_facility_name(providers)
-
-        providers
       end
 
       def merge_fields(_options = {})
         expand_va_file_number
 
         expand_ssn
-        expand_phone_number
 
         expand_veteran_full_name
-        expand_printed_full_name
-
         signature_date = @form_data['signatureDate']
         expand_signature(@form_data['veteranFullName'], signature_date)
-        expand_signature_date
+        @form_data['printedName'] = @form_data['signature']
         @form_data['signature'] = "/es/ #{@form_data['signature']}"
 
         expand_claimant_address
@@ -602,293 +365,9 @@ binding.pry
         expand_veteran_service_number
 
         @form_data['providerFacility'] = expand_providers(@form_data['providerFacility'])
-binding.pry
+
         @form_data
       end
     end
   end
 end
-
-# Source of truth: pdftk output fields
-# F[0].Page_1[0].VeteranFirstName[0]
-# F[0].Page_1[0].VeteranMiddleInitial1[0]
-# F[0].Page_1[0].VeteranLastName[0]
-# F[0].Page_1[0].VeteransSocialSecurityNumber_FirstThreeNumbers[0]
-# F[0].Page_1[0].VeteransSocialSecurityNumber_SecondTwoNumbers[0]
-# F[0].Page_1[0].VeteransSocialSecurityNumber_LastFourNumbers[0]
-# F[0].Page_1[0].VAFileNumber[0]
-# F[0].Page_1[0].DOBmonth[0]
-# F[0].Page_1[0].DOBday[0]
-# F[0].Page_1[0].DOByear[0]
-# F[0].Page_1[0].VeteransServiceNumber_If_Applicable[0]
-# F[0].Page_1[0].MailingAddress_NumberAndStreet[0]
-# F[0].Page_1[0].MailingAddress_ApartmentOrUnitNumber[0]
-# F[0].Page_1[0].MailingAddress_City[0]
-# F[0].Page_1[0].MailingAddress_StateOrProvince[0]
-# F[0].Page_1[0].MailingAddress_Country[0]
-# F[0].Page_1[0].MailingAddress_ZIPOrPostalCode_FirstFiveNumbers[0]
-# F[0].Page_1[0].MailingAddress_ZIPOrPostalCode_LastFourNumbers[0]
-# F[0].Page_1[0].TelephoneNumber_AreaCode[0]
-# F[0].Page_1[0].TelephoneNumber_SecondThreeNumbers[0]
-# F[0].Page_1[0].TelephoneNumber_LastFourNumbers[0]
-# F[0].Page_1[0].International_Telephone_Number_If_Applicable[0]
-# F[0].Page_1[0].CheckBox1[0]
-# F[0].Page_1[0].E_Mail_Address[0]
-# F[0].Page_1[0].E_Mail_Address[1]
-# F[0].Page_1[0].Patients_FirstName[0]
-# F[0].Page_1[0].Patients_MiddleInitial1[0]
-# F[0].Page_1[0].Patients_LastName[0]
-# F[0].Page_1[0].Patient_SocialSecurityNumber_FirstThreeNumbers[0]
-# F[0].Page_1[0].Patient_SocialSecurityNumber_SecondTwoNumbers[0]
-# F[0].Page_1[0].Patient_SocialSecurityNumber_LastFourNumbers[0]
-# F[0].Page_1[0].Patients_VAFileNumber_If_Applicable[0]
-# F[0].#subform[1].VeteransSocialSecurityNumber_FirstThreeNumbers[0]
-# F[0].#subform[1].VeteransSocialSecurityNumber_SecondTwoNumbers[0]
-# F[0].#subform[1].VeteransSocialSecurityNumber_LastFourNumbers[0]
-# F[0].#subform[1].InformationIsLimitedToWhatIsWrittenInThisSpace[0]
-# F[0].#subform[1].SignatureField11[0]
-# F[0].#subform[1].Date_Signed_Month[0]
-# F[0].#subform[1].Date_Signed_Day[0]
-# F[0].#subform[1].Date_Signed_Year[0]
-# F[0].#subform[1].Printed_Name_Of_Person_Signing_First[0]
-# F[0].#subform[1].Printed_Name_Of_Person_Signing_Middle_Initial[0]
-# F[0].#subform[1].Printed_Name_Of_Person_Signing_Last[0]
-# F[0].#subform[1].Relationship_To_Veteran_Claimant[0]
-# F[0].#subform[14].VeteranFirstName[0]
-# F[0].#subform[14].VeteranMiddleInitial1[0]
-# F[0].#subform[14].VeteranLastName[0]
-# F[0].#subform[14].SSN1[0]
-# F[0].#subform[14].SSN2[0]
-# F[0].#subform[14].SSN3[0]
-# F[0].#subform[14].VAFileNumber[0]
-# F[0].#subform[14].Month[0]
-# F[0].#subform[14].Day[0]
-# F[0].#subform[14].Year[0]
-# F[0].#subform[14].VeteransServiceNumber_If_Applicable[0]
-# F[0].#subform[14].Patients_FirstName[0]
-# F[0].#subform[14].PatientMiddleInitial1[0]
-# F[0].#subform[14].Patients_LastName[0]
-# F[0].#subform[14].FirstThreeNumbers[0]
-# F[0].#subform[14].SecondTwoNumbers[0]
-# F[0].#subform[14].LastFourNumbers[0]
-# F[0].#subform[14].VAFileNumber[1]
-# F[0].#subform[14].Provider_Or_Facility_Name[0]
-# F[0].#subform[14].Conditions_You_Are_Being_Treated_For[0]
-# F[0].#subform[14].Month[1]
-# F[0].#subform[14].Day[1]
-# F[0].#subform[14].Year[1]
-# F[0].#subform[14].Month[2]
-# F[0].#subform[14].Day[2]
-# F[0].#subform[14].Year[2]
-# F[0].#subform[14].Provider_Facility_Street_Address_NumberAndStreet[0]
-# F[0].#subform[14].MailingAddress_ApartmentOrUnitNumber[0]
-# F[0].#subform[14].Provider_Facility_Address_City[0]
-# F[0].#subform[14].Provider_Facility_Address_StateOrProvince[0]
-# F[0].#subform[14].Provider_Facility_Address_Country[0]
-# F[0].#subform[14].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[0]
-# F[0].#subform[14].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[0]
-# F[0].#subform[14].Provider_Or_Facility_Name[1]
-# F[0].#subform[14].Conditions_You_Are_Being_Treated_For[1]
-# F[0].#subform[14].Month[3]
-# F[0].#subform[14].Day[3]
-# F[0].#subform[14].Year[3]
-# F[0].#subform[14].Month[4]
-# F[0].#subform[14].Day[4]
-# F[0].#subform[14].Year[4]
-# F[0].#subform[14].Provider_Facility_Street_Address_NumberAndStreet[1]
-# F[0].#subform[14].MailingAddress_ApartmentOrUnitNumber[1]
-# F[0].#subform[14].Provider_Facility_Address_City[1]
-# F[0].#subform[14].Provider_Facility_Address_StateOrProvince[1]
-# F[0].#subform[14].Provider_Facility_Address_Country[1]
-# F[0].#subform[14].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[1]
-# F[0].#subform[14].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[1]
-# F[0].#subform[15].SSN1[1]
-# F[0].#subform[15].SSN2[1]
-# F[0].#subform[15].SSN3[1]
-# F[0].#subform[15].Provider_Or_Facility_Name[2]
-# F[0].#subform[15].Conditions_You_Are_Being_Treated_For[2]
-# F[0].#subform[15].Month[5]
-# F[0].#subform[15].Day[5]
-# F[0].#subform[15].Year[5]
-# F[0].#subform[15].Month[6]
-# F[0].#subform[15].Day[6]
-# F[0].#subform[15].Year[6]
-# F[0].#subform[15].Provider_Facility_Street_Address_NumberAndStreet[2]
-# F[0].#subform[15].MailingAddress_ApartmentOrUnitNumber[2]
-# F[0].#subform[15].Provider_Facility_Address_City[2]
-# F[0].#subform[15].Provider_Facility_Address_StateOrProvince[2]
-# F[0].#subform[15].Provider_Facility_Address_Country[2]
-# F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[2]
-# F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[2]
-# F[0].#subform[15].Provider_Or_Facility_Name[3]
-# F[0].#subform[15].Conditions_You_Are_Being_Treated_For[3]
-# F[0].#subform[15].Month[7]
-# F[0].#subform[15].Day[7]
-# F[0].#subform[15].Year[7]
-# F[0].#subform[15].Month[8]
-# F[0].#subform[15].Day[8]
-# F[0].#subform[15].Year[8]
-# F[0].#subform[15].Provider_Facility_Street_Address_NumberAndStreet[3]
-# F[0].#subform[15].MailingAddress_ApartmentOrUnitNumber[3]
-# F[0].#subform[15].Provider_Facility_Address_City[3]
-# F[0].#subform[15].Provider_Facility_Address_StateOrProvince[3]
-# F[0].#subform[15].Provider_Facility_Address_Country[3]
-# F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[3]
-# F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[3]
-# F[0].#subform[15].Provider_Or_Facility_Name[4]
-# F[0].#subform[15].Conditions_You_Are_Being_Treated_For[4]
-# F[0].#subform[15].Month[9]
-# F[0].#subform[15].Day[9]
-# F[0].#subform[15].Year[9]
-# F[0].#subform[15].Month[10]
-# F[0].#subform[15].Day[10]
-# F[0].#subform[15].Year[10]
-# F[0].#subform[15].Provider_Facility_Street_Address_NumberAndStreet[4]
-# F[0].#subform[15].MailingAddress_ApartmentOrUnitNumber[4]
-# F[0].#subform[15].Provider_Facility_Address_City[4]
-# F[0].#subform[15].Provider_Facility_Address_StateOrProvince[4]
-# F[0].#subform[15].Provider_Facility_Address_Country[4]
-# F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[4]
-# F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[4]
-
-# 4142 standalone hash being passed to pdftk to populate the form fields
-
-# "F[0].Page_1[0].VeteranFirstName[0]"=>"Hector",
-# "F[0].Page_1[0].VeteranMiddleInitial1[0]"=>"J ",
-# "F[0].Page_1[0].VeteranLastName[0]"=>"Allen",
-# "F[0].Page_1[0].VeteransSocialSecurityNumber_FirstThreeNumbers[0]"=>"123",
-# "F[0].Page_1[0].VeteransSocialSecurityNumber_SecondTwoNumbers[0]"=>"49",
-# "F[0].Page_1[0].VeteransSocialSecurityNumber_LastFourNumbers[0]"=>"8762",
-# "F[0].Page_1[0].VAFileNumber[0]"=>"",
-# "F[0].Page_1[0].DOByear[0]"=>"1996",
-# "F[0].Page_1[0].DOBmonth[0]"=>"04",
-# "F[0].Page_1[0].DOBday[0]"=>"13",
-# "F[0].Page_1[0].VeteransServiceNumber_If_Applicable[0]"=>"",
-# "F[0].Page_1[0].MailingAddress_ZIPOrPostalCode_FirstFiveNumbers[0]"=>"90806",
-# "F[0].Page_1[0].MailingAddress_ZIPOrPostalCode_LastFourNumbers[0]"=>"",
-# "F[0].Page_1[0].MailingAddress_Country[0]"=>"USA",
-# "F[0].Page_1[0].MailingAddress_StateOrProvince[0]"=>"CA",
-# "F[0].Page_1[0].MailingAddress_City[0]"=>"Los Angeles",
-# "F[0].Page_1[0].MailingAddress_ApartmentOrUnitNumber[0]"=>"",
-# "F[0].Page_1[0].MailingAddress_NumberAndStreet[0]"=>"1234 Fake st ",
-# "F[0].Page_1[0].TelephoneNumber_AreaCode[0]"=>"123",
-# "F[0].Page_1[0].TelephoneNumber_SecondThreeNumbers[0]"=>"876",
-# "F[0].Page_1[0].TelephoneNumber_LastFourNumbers[0]"=>"2315",
-# "F[0].Page_1[0].International_Telephone_Number_If_Applicable[0]"=>"",
-# "F[0].Page_1[0].CheckBox1[0]"=>"1",
-# "F[0].Page_1[0].E_Mail_Address[0]"=>"Fake@email.com",
-# "F[0].Page_1[0].E_Mail_Address[1]"=>"",
-# "F[0].Page_1[0].Patients_FirstName[0]"=>"Hector",
-# "F[0].Page_1[0].Patients_MiddleInitial1[0]"=>"J",
-# "F[0].Page_1[0].Patients_LastName[0]"=>"Allen",
-# "F[0].Page_1[0].Patient_SocialSecurityNumber_FirstThreeNumbers[0]"=>"897",
-# "F[0].Page_1[0].Patient_SocialSecurityNumber_SecondTwoNumbers[0]"=>"12",
-# "F[0].Page_1[0].Patient_SocialSecurityNumber_LastFourNumbers[0]"=>"3645",
-# "F[0].Page_1[0].Patients_VAFileNumber_If_Applicable[0]"=>"",
-# "F[0].#subform[1].VeteransSocialSecurityNumber_FirstThreeNumbers[0]"=>"123",
-# "F[0].#subform[1].VeteransSocialSecurityNumber_SecondTwoNumbers[0]"=>"49",
-# "F[0].#subform[1].VeteransSocialSecurityNumber_LastFourNumbers[0]"=>"8762",
-# "F[0].#subform[1].InformationIsLimitedToWhatIsWrittenInThisSpace[0]"=>"",
-# "F[0].#subform[1].SignatureField11[0]"=>"Hector J  Allen",
-# "F[0].#subform[1].Date_Signed_Year[0]"=>"2025",
-# "F[0].#subform[1].Date_Signed_Month[0]"=>"05",
-# "F[0].#subform[1].Date_Signed_Day[0]"=>"13",
-# "F[0].#subform[1].Printed_Name_Of_Person_Signing_First[0]"=>"",
-# "F[0].#subform[1].Printed_Name_Of_Person_Signing_Middle_Initial[0]"=>"",
-# "F[0].#subform[1].Printed_Name_Of_Person_Signing_Last[0]"=>"",
-# "F[0].#subform[1].Relationship_To_Veteran_Claimant[0]"=>"I am the Veteran           ",
-# "F[0].#subform[14].VeteranFirstName[0]"=>"Hector",
-# "F[0].#subform[14].VeteranMiddleInitial1[0]"=>"J ",
-# "F[0].#subform[14].VeteranLastName[0]"=>"Allen",
-# "F[0].#subform[14].SSN1[0]"=>"123",
-# "F[0].#subform[14].SSN2[0]"=>"49",
-# "F[0].#subform[14].SSN3[0]"=>"8762",
-# "F[0].#subform[14].VAFileNumber[0]"=>"",
-# "F[0].#subform[14].Year[0]"=>"1996",
-# "F[0].#subform[14].Month[0]"=>"04",
-# "F[0].#subform[14].Day[0]"=>"13",
-# "F[0].#subform[14].VeteransServiceNumber_If_Applicable[0]"=>"",
-# "F[0].#subform[14].Patients_FirstName[0]"=>"Hector",
-# "F[0].#subform[14].PatientMiddleInitial1[0]"=>"J",
-# "F[0].#subform[14].Patients_LastName[0]"=>"Allen",
-# "F[0].#subform[14].FirstThreeNumbers[0]"=>"897",
-# "F[0].#subform[14].SecondTwoNumbers[0]"=>"12",
-# "F[0].#subform[14].LastFourNumbers[0]"=>"3645",
-# "F[0].#subform[14].VAFileNumber[1]"=>"",
-# "F[0].#subform[14].Provider_Or_Facility_Name[0]"=>"VA",
-# "F[0].#subform[14].Conditions_You_Are_Being_Treated_For[0]"=>"PTSD",
-# "F[0].#subform[14].Year[1]"=>"2020",
-# "F[0].#subform[14].Month[1]"=>"06",
-# "F[0].#subform[14].Day[1]"=>"12",
-# "F[0].#subform[14].Year[2]"=>"2020",
-# "F[0].#subform[14].Month[2]"=>"12",
-# "F[0].#subform[14].Day[2]"=>"15",
-# "F[0].#subform[14].Provider_Facility_Street_Address_NumberAndStreet[0]"=>"1234 Fake St",
-# "F[0].#subform[14].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[0]"=>"90806",
-# "F[0].#subform[14].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[0]"=>"",
-# "F[0].#subform[14].Provider_Facility_Address_Country[0]"=>"USA",
-# "F[0].#subform[14].Provider_Facility_Address_StateOrProvince[0]"=>"CA",
-# "F[0].#subform[14].Provider_Facility_Address_City[0]"=>"Long Beach",
-# "F[0].#subform[14].Provider_Or_Facility_Name[1]"=>"",
-# "F[0].#subform[14].Conditions_You_Are_Being_Treated_For[1]"=>"",
-# "F[0].#subform[14].Year[3]"=>"",
-# "F[0].#subform[14].Month[3]"=>"",
-# "F[0].#subform[14].Day[3]"=>"",
-# "F[0].#subform[14].Year[4]"=>"",
-# "F[0].#subform[14].Month[4]"=>"",
-# "F[0].#subform[14].Day[4]"=>"",
-# "F[0].#subform[14].Provider_Facility_Street_Address_NumberAndStreet[1]"=>"",
-# "F[0].#subform[14].MailingAddress_ApartmentOrUnitNumber[1]"=>"",
-# "F[0].#subform[14].Provider_Facility_Address_City[1]"=>"",
-# "F[0].#subform[14].Provider_Facility_Address_StateOrProvince[1]"=>"",
-# "F[0].#subform[14].Provider_Facility_Address_Country[1]"=>"",
-# "F[0].#subform[14].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[1]"=>"",
-# "F[0].#subform[14].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[1]"=>"",
-# "F[0].#subform[15].SSN1[1]"=>"123",
-# "F[0].#subform[15].SSN2[1]"=>"49",
-# "F[0].#subform[15].SSN3[1]"=>"8762",
-# "F[0].#subform[15].Provider_Or_Facility_Name[2]"=>"",
-# "F[0].#subform[15].Conditions_You_Are_Being_Treated_For[2]"=>"",
-# "F[0].#subform[15].Year[5]"=>"",
-# "F[0].#subform[15].Month[5]"=>"",
-# "F[0].#subform[15].Day[5]"=>"",
-# "F[0].#subform[15].Year[6]"=>"",
-# "F[0].#subform[15].Month[6]"=>"",
-# "F[0].#subform[15].Day[6]"=>"",
-# "F[0].#subform[15].Provider_Facility_Street_Address_NumberAndStreet[2]"=>"",
-# "F[0].#subform[15].MailingAddress_ApartmentOrUnitNumber[2]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_City[2]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_StateOrProvince[2]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_Country[2]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[2]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[2]"=>"",
-# "F[0].#subform[15].Provider_Or_Facility_Name[3]"=>"",
-# "F[0].#subform[15].Conditions_You_Are_Being_Treated_For[3]"=>"",
-# "F[0].#subform[15].Year[7]"=>"",
-# "F[0].#subform[15].Month[7]"=>"",
-# "F[0].#subform[15].Day[7]"=>"",
-# "F[0].#subform[15].Year[8]"=>"",
-# "F[0].#subform[15].Month[8]"=>"",
-# "F[0].#subform[15].Day[8]"=>"",
-# "F[0].#subform[15].Provider_Facility_Street_Address_NumberAndStreet[3]"=>"",
-# "F[0].#subform[15].MailingAddress_ApartmentOrUnitNumber[3]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_City[3]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_StateOrProvince[3]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_Country[3]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[3]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[3]"=>"",
-# "F[0].#subform[15].Provider_Or_Facility_Name[4]"=>"",
-# "F[0].#subform[15].Conditions_You_Are_Being_Treated_For[4]"=>"",
-# "F[0].#subform[15].Year[9]"=>"",
-# "F[0].#subform[15].Month[9]"=>"",
-# "F[0].#subform[15].Day[9]"=>"",
-# "F[0].#subform[15].Year[10]"=>"",
-# "F[0].#subform[15].Month[10]"=>"",
-# "F[0].#subform[15].Day[10]"=>"",
-# "F[0].#subform[15].Provider_Facility_Street_Address_NumberAndStreet[4]"=>"",
-# "F[0].#subform[15].MailingAddress_ApartmentOrUnitNumber[4]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_City[4]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_StateOrProvince[4]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_Country[4]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_FirstFiveNumbers[4]"=>"",
-# "F[0].#subform[15].Provider_Facility_Address_ZIPOrPostalCode_LastFourNumbers[4]"=>""}
